完成ItemBasics
