+++ rn/components/Item/index.js
@@ -1,35 +1,11 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { Text, View, Image, TouchableOpacity } from 'react-native';
+import ItemBasics from '../ItemBasics';
 
-export default class Item extends Component {
+export default class Item extends ItemBasics {
   static propTypes = {
-    leftIcon: PropTypes.oneOfType([
-      PropTypes.element,
-      PropTypes.shape({ uri: PropTypes.string }),
-      PropTypes.number,
-    ]),
-    rightIcon: PropTypes.oneOfType([
-      PropTypes.element,
-      PropTypes.shape({ uri: PropTypes.string }),
-      PropTypes.number,
-      PropTypes.oneOf(['none', 'empty', 'check', 'indicator']),
-    ]),
-    title: PropTypes.oneOfType([
-      PropTypes.element,
-      PropTypes.string,
-      PropTypes.number,
-    ]).isRequired,
-    note: PropTypes.oneOfType([
-      PropTypes.element,
-      PropTypes.string,
-      PropTypes.number,
-    ]),
-    details: PropTypes.oneOfType([
-      PropTypes.element,
-      PropTypes.string,
-      PropTypes.number,
-    ]),
+    ...ItemBasics.propTypes,
   };
 
   static defaultProps = {
@@ -156,38 +132,5 @@ export default class Item extends Component {
       );
     }
     return note;
-  }
-
-  render() {
-    let {
-      leftIcon,
-      rightIcon,
-      title,
-      note,
-      details,
-      children,
-      ...other
-    } = this.props;
-    const leftIconCont = this.renderLeftIcon(leftIcon);
-    const rightIconCont = this.renderRightIcon(rightIcon);
-    const titleCont = this.renderTitle(title, details);
-    const noteCont = this.renderNote(note);
-    const contentStyle = {
-      overflow: 'hidden',
-      flexDirection: 'row',
-      alignItems: 'center',
-      justifyContent: 'space-between',
-    };
-    return (
-      <TouchableOpacity {...other}>
-        <View style={contentStyle}>
-          {leftIconCont}
-          {titleCont}
-          {children && children}
-          {!children && noteCont}
-          {!children && rightIconCont}
-        </View>
-      </TouchableOpacity>
-    );
   }
 }
